pyproject.toml
[tool.poetry]
name = "zik-discord"
version = "0.1.0"
description = ""
authors = ["Zik <vladziku@gmail.com>"]
readme = "README.md"
packages = [{include = "zik_discord"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
